Enable APCu
PHP 5.5 is shipped with Zend OPcache, which supersedes a lot of APC
features. Yet, APC has an user cache, not exposed in the Zend OPcache
public API. APCu is an extension containing this user cache from APC.

Currently, APCu is still offered in beta state, so we tweak the PEAR
distribution channel to accept beta packages.