Essential installation tasks at first run
Currently, our image has been optimized to run an already installed
instance of Phabricator, or install it manually.

Yet, it would be more convenient if when we start a new container
for the first time, we've a working installation.

The setup-phabricator script takes care:
    * to populate database or upgrade schema

We so introduce a new environment variable, PHABRICATOR_NO_INSTALL.
If set, this bypass this part of the script. You should set it if
you deploy your own Phabricator code instead of the master branch,
and you don't systematically rebase against master, to avoid to get
an upgraded MySQL schema.