Essential installation tasks at first run
Summary:
Currently, our image has been optimized to run an already installed
instance of Phabricator, or install it manually.

Yet, it would be more convenient if when we start a new container
for the first time, we've a working installation.

The setup-phabricator script takes care:
    * to populate database or upgrade schema

We so introduce a new environment variable, PHABRICATOR_NO_INSTALL.
If set, this bypass this part of the script. You should set it if
you deploy your own Phabricator code instead of the master branch,
and you don't systematically rebase against master, to avoid to get
an upgraded MySQL schema.

Depends on D40 for the documentation part.

Test Plan: P124

Reviewers: Kaliiixx, dereckson

Projects: #nasqueron_docker_deployment_squad

Maniphest Tasks: T329

Differential Revision: http://devcentral.nasqueron.org/D44