Allow to set MySQL credentials
Summary:
If the DB_USER environment variable is set, instead to use the root account
of the linked MySQL container, Phabricator will be configured to use both
DB_USER and DB_PASS values to connect to the database.

This is a part of the migration from deprecated Docker links to networks.

Test Plan: Test with Zed Phabricator instance

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: #nasqueron_docker_deployment_squad

Tags: #docker_images

Differential Revision: https://devcentral.nasqueron.org/D1917